some numpy dtype fixes, other minor cleanup

+++ Lib/sandbox/models/utils.py
@@ -10,7 +10,7 @@ def recipr(X):
     equal to 0 to 0. Therefore, it presumes that X should be positive in
     general.
     """
-    x = N.maximum(N.asarray(X).astype(N.Float), 0)
+    x = N.maximum(N.asarray(X).astype(N.float64), 0)
     return N.greater(x, 0.) / (x + N.less_equal(x, 0.))
 
 def mad(a, c=0.6745, axis=0):
@@ -52,7 +52,7 @@ def rank(X, cond=1.0e-12):
     not the SVD.
     """
     D = scipy.linalg.svdvals(X)
-    return int(N.add.reduce(N.greater(D / D.max(), cond).astype(N.Int)))
+    return int(N.add.reduce(N.greater(D / D.max(), cond).astype(N.int32)))
 
 def fullrank(X, r=None):
     """
@@ -72,7 +72,7 @@ def fullrank(X, r=None):
     value = []
     for i in range(r):
         value.append(V[:,order[i]])
-    return N.asarray(N.transpose(value)).astype(N.Float)
+    return N.asarray(N.transpose(value)).astype(N.float64)
 
 class StepFunction:
     '''A basic step function: values at the ends are handled in the simplest way possible: everything to the left of x[0] is set to ival; everything to the right of x[-1] is set to y[-1].
